Fixing some compiler warnings (unused variables, etc.).
git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3742 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/src/gdevepsn.c
@@ -409,7 +409,7 @@ eps_output_run(byte *data, int count, int y_mult,
 	fputc(033, prn_stream);
 	if ( !(start_graphics & ~3) )
 	{	
-		fputc("KLYZ"[start_graphics], prn_stream);
+		fputc("KLYZ"[(int)start_graphics], prn_stream);
 	}
 	else
 	{
